refactor : fixture와 fake 객체 패키지 변경 (#54)
