+++ src/test/java/gdsc/binaryho/imhere/mock/fakerepository/FakeAttendeeCacheRepository.java
@@ -1,4 +1,4 @@
-package gdsc.binaryho.imhere.mock;
+package gdsc.binaryho.imhere.mock.fakerepository;
 
 import gdsc.binaryho.imhere.core.lecture.application.port.AttendeeCacheRepository;
 import gdsc.binaryho.imhere.core.lecture.model.StudentIds;
