+++ src/test/java/gdsc/binaryho/imhere/mock/fixture/MemberFixture.java
@@ -1,4 +1,4 @@
-package gdsc.binaryho.imhere.fixture;
+package gdsc.binaryho.imhere.mock.fixture;
 
 import gdsc.binaryho.imhere.core.member.Member;
 import gdsc.binaryho.imhere.core.member.Role;
@@ -11,7 +11,6 @@ public class MemberFixture {
     public static final String RAW_PASSWORD = "abcd1234";
     public static final String PASSWORD =
         new BCryptPasswordEncoder().encode(RAW_PASSWORD);
-    public static final Role ROLE = Role.STUDENT;
 
     public static final Long STUDENT_ID = 1L;
     public static final Long LECTURER_ID = 2L;
